% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PvalToStar.R
\name{PvalToStar}
\alias{PvalToStar}
\title{Assign significance stars to p-values}
\usage{
PvalToStar(p.val, sig.levels = c(0.01, 0.05, 0.1), sig.symbols = c("***",
  "**", "*"))
}
\arguments{
\item{p.val}{Vector of p-values to which we want significance stars assigned.}

\item{sig.levels}{Optional vector containing significance thresholds. Default is 1\%, 5\%, and 10\% significance.}

\item{sig.symbols}{Vector of symbols to denote significance. Symbols must be provided in decreasing order of significance.}
}
\value{
A vector of significance stars.
}
\description{
Assign significance stars to p-values
}
\examples{
PvalToStar(c(0.009, 0.049, 0.09, 0.11))

my.levels <- c(0.001, 0.01, 0.05)
my.symbols <- c("a", "b", "+")
PvalToStar(c(0.009, 0.049, 0.09, 0.11), my.levels, my.symbols)
}
