% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegToDT.R
\name{RegToDT}
\alias{RegToDT}
\title{Output Regression Results to data.table}
\usage{
RegToDT(reg.model, digits = 3, print.tstat = FALSE, sig.levels = c(0.01,
  0.05, 0.1), sig.symbols = c("***", "**", "*"))
}
\arguments{
\item{reg.model}{Regression model, e.g. an lm object.}

\item{digits}{Number of digits to round coefficients, standard errors.}

\item{print.tstat}{Should t-stats be displayed instead of standard errors?}

\item{sig.levels}{Significance levels based on which stars should be assigned, if any.}

\item{sig.symbols}{Significance symbols to assign. Should be in descending order of significance.}
}
\value{
A data.table of regression coefficients, standard errors (or t-stats), R-squared, number of observations. Standard errors (or t-stats) are labelled starting with "_SE_". If available (e.g. for an felm object), the R-squared of the projected model is displayed, or, if a 2SLS model is estimated via felm, the F-statistic of the first stage (instead of R-squareds in this case).
}
\description{
Output Regression Results to data.table
}
\examples{
reg.results <- lm(mpg ~ disp + hp, mtcars)
RegToDT(reg.results)
}
