% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRDSConnect.R
\name{WRDSConnect}
\alias{WRDSConnect}
\title{Set up connection to WRDS}
\usage{
WRDSConnect(user, pass, jar.path)
}
\arguments{
\item{user}{WRDS user name.}

\item{pass}{WRDS password (recommended to use hashed version).}

\item{jar.path}{Local path to the necessary .jar file (available on WRDS).}
}
\value{
Connection to WRDS
}
\description{
Follows the instructions provided by WRDS.
}
