% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRDSDownload.R
\name{WRDSDownload}
\alias{WRDSDownload}
\title{Download data from WRDS}
\usage{
WRDSDownload(qstr, nobs = -1, con = wrds, out.dir = data.dir,
  dt.format = TRUE)
}
\arguments{
\item{qstr}{A vector of three strings: (i) name of source dataset, (ii) query for data to be pulled, (iii) query for variable names or supplied variable names.}

\item{nobs}{Number of observations to download (default is all).}

\item{con}{Name of the connection (default wrds).}

\item{out.dir}{Path to directory for saving downloaded dataset.}

\item{dt.format}{Should the saved dataset be in data.table format? Default is TRUE.}
}
\description{
Makes a query for the WRDS server, and saves the resulting dataset to disk.
}
